refactor:entity 패키지 이름을 domain으로 수정
