+++ backend/src/main/java/com/graphy/backend/domain/member/domain/Role.java
@@ -1,0 +1,6 @@
+package com.graphy.backend.domain.member.domain;
+
+
+public enum Role {
+    ROLE_USER
+}
