--- backend/src/main/java/com/graphy/backend/domain/member/entity/Role.java
@@ -1,6 +1,0 @@
-package com.graphy.backend.domain.member.entity;
-
-
-public enum Role {
-    ROLE_USER
-}
